BabySqueel::Table is now BabySqueel::Relation. Table should be simpler and not require an activerecord model, just a traditional Arel::Table
