+++ lib/baby_squeel.rb
@@ -12,6 +12,18 @@ module BabySqueel
     require 'baby_squeel/compat'
     BabySqueel::Compat.enable!
   end
+
+  class NotFoundError < StandardError
+    def initialize(model_name, name)
+      super "There is no column or association named '#{name}' for #{model_name}."
+    end
+  end
+
+  class AssociationNotFoundError < StandardError
+    def initialize(model_name, name)
+      super "Association named '#{name}' was not found for #{model_name}."
+    end
+  end
 end
 
 ::ActiveRecord::Base.extend BabySqueel::ActiveRecord::Sifting
